+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add letter based SIDC to number SIDC converter to symbol picker
 - Grid cell copy/paste
 - Grid edit mode keyboard navigation